attempt 1, this works but seems odd.
