+++ src/DynamoCore/Logging/DynamoAnalyticsClient.cs
@@ -12,7 +12,6 @@ namespace Dynamo.Logging
 {
     class DynamoAnalyticsSession : IAnalyticsSession
     {
-        private Heartbeat heartbeat;
         private UsageLog logger;
 
         public DynamoAnalyticsSession()
@@ -25,8 +24,6 @@ namespace Dynamo.Logging
         {
             StabilityCookie.Startup();
 
-            heartbeat = Heartbeat.GetInstance(model);
-
             logger = new UsageLog("Dynamo", UserId, SessionId);
         }
 
@@ -37,10 +34,6 @@ namespace Dynamo.Logging
                 StabilityCookie.WriteCrashingShutdown();
             else
                 StabilityCookie.WriteCleanShutdown();
-
-            if (null != heartbeat)
-                Heartbeat.DestroyInstance();
-            heartbeat = null;
 
             if (null != logger)
                 logger.Dispose();
@@ -178,7 +171,7 @@ namespace Dynamo.Logging
 
             if (Session == null) Session = new DynamoAnalyticsSession();
 
-            //Setup Analytics service, StabilityCookie, Heartbeat and UsageLog.
+            //Setup Analytics service, StabilityCookie, and UsageLog.
             Session.Start(dynamoModel);
 
             //Dynamo app version.
